Move files into folders. Add index files. Remove css files because we will be using styled-components.
